requirements.txt
Flask
pytesseract
opencv-python
Pillow
python-dotenv
flask-cors